+++ CHANGELOG.md
@@ -4,8 +4,9 @@ Notable changes to this project will be documented in this file.
 
 ## [v2.0.3]
 
-- Cleanup of legacy/duplicate clones that used a different `zstyle ':antidote:bundle' path-style`
+- Cleanup of legacy/duplicate clones that used a different `zstyle ':antidote:bundle' path-style` during bundling
 - Fix antidote list [#245](https://github.com/mattmc3/antidote/issues/245)
+- Fix for find bundles when `zstyle ':antidote:bundle' path-style escaped`, but was not previously and bundles were cloned using a different path style
 
 ## [v2.0.2]