+++ CHANGELOG.md
@@ -4,9 +4,11 @@ Notable changes to this project will be documented in this file.
 
 ## [v2.0.3]
 
-- Cleanup of legacy/duplicate clones that used a different `zstyle ':antidote:bundle' path-style` during bundling
-- Fix antidote list [#245](https://github.com/mattmc3/antidote/issues/245)
-- Fix for find bundles when `zstyle ':antidote:bundle' path-style escaped`, but was not previously and bundles were cloned using a different path style
+- Reuse existing clones when `path-style` changes, avoiding duplicate clones ([#245](https://github.com/mattmc3/antidote/issues/245))
+- Remove legacy duplicate clones during bundling when multiple path-style directories exist
+- Fix `find_bundles` failing when cloned bundles don't match the current `path-style`
+- Fix `antidote bundle` emitting output before ensuring a successful clone operation
+- `antidote list` now shows path and URL by default, and now has a `-u/--url` flag
 
 ## [v2.0.2]